Fixes #336.  Don't add 'NSS-<id>' to ref list when importing Nessus nbe/xml to database (report/patch by erwinpdev).
This NSS- stuff was in the original nbe code before things were copied and
reworked for the xml handling.  I'm not sure why it was there; the NSS id is
already available in the database (name).

Note that ref lists generated here can be empty now, which I think is fine.



git-svn-id: https://www.metasploit.com/svn/framework3/trunk@7040 4d416f70-5f16-0410-b530-b9f4589650da